+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Master
 
+## v128 (11/4/2014)
+
+* Better cedar14 Ruby install error message
+
 ## v127 (9/18/2014)
 
 * rbx is now stack aware